Use pkg-config to locate mysqlclient library
There are some pitfalls for the mysql_config approach. It actually
outputs the unparsed CFLAGS & LDFLAGS. The include path generated by
cmake is concatenated without spaces and is not valid. The build will
fail if the header is not installed to standard location. The LDFLAGS
also includes some private libraries like ssl & crypto, which are not
required. This commit uses the cmake pkg-config module to generate the
proper variables and fixes the mysql.h header inclusion.