Optimize job maker ZooKeeper lock
There is a ZooKeeper lock in job maker to prevent multiple instances of
the same job maker configuration. This is to ensure the job id
uniqueness. However we can embed the server id in job id to ensure this
while running multiple job maker instances. The ZooKeeper lock now locks
the path with server id at the end so that we won't start the job maker
with the same server id.