ETH: use forked eth_getWork()
ETH RPC eth_getBlockByNumber() will not return the correct pending block
for the very first call of a specifc height as geth is not populating
the information yet (at that time only a work package for an empty block
is created). This leads to incorrect gas/transaction/uncle in the
gwmaker being populated. Hence we forked geth & parity to get all
required information in eth_getWork() result.